Improve the GRDB demo.
