+++ Examples/GRDBDemo/Schema.swift
@@ -20,6 +20,18 @@ extension Player: Codable, FetchableRecord, MutablePersistableRecord {
 extension DatabaseWriter {
   func migrate() throws {
     var migrator = DatabaseMigrator()
+    defer {
+      #if targetEnvironment(simulator)
+        migrator.registerMigration("Create seed data") { db in
+          try Player.deleteAll(db)
+          for (index, name) in ["Blob", "Blob Jr.", "Blob Sr.", "Blob Esq."].enumerated() {
+            _ = try Player(name: name, isInjured: index.isMultiple(of: 2))
+              .inserted(db)
+          }
+        }
+      #endif
+      try! migrator.migrate(self)
+    }
     #if DEBUG
       migrator.eraseDatabaseOnSchemaChange = true
     #endif
@@ -30,15 +42,6 @@ extension DatabaseWriter {
         t.column("isInjured", .boolean).defaults(to: false).notNull()
       }
     }
-    #if targetEnvironment(simulator)
-    migrator.registerMigration("Create seed data") { db in
-      for (index, name) in ["Blob", "Blob Jr.", "Blob Sr.", "Blob Esq."].enumerated() {
-        _ = try Player(name: name, isInjured: index.isMultiple(of: 2))
-          .inserted(db)
-      }
-    }
-    #endif
-    try migrator.migrate(self)
   }
 }
 
@@ -56,48 +59,17 @@ extension SharedReaderKey where Self == GRDBQueryKey<PlayersRequest>.Default {
 struct PlayersRequest: GRDBQuery {
   let order: PlayerOrder
   func fetch(_ db: Database) throws -> [Player] {
-    let ordering: any SQLOrderingTerm = switch order {
-    case .name:
-      Column("name")
-    case .isInjured:
-      Column("isInjured").desc
-    }
-    return try Player
+    let ordering: any SQLOrderingTerm =
+      switch order {
+      case .name:
+        Column("name")
+      case .isInjured:
+        Column("isInjured").desc
+      }
+    return
+      try Player
       .all()
       .order(ordering)
       .fetchAll(db)
-  }
-}
-
-extension DependencyValues {
-  public var database: DatabaseQueue {
-    get { self[GRDBDatabaseKey.self] }
-    set { self[GRDBDatabaseKey.self] = newValue }
-  }
-}
-
-private enum GRDBDatabaseKey: DependencyKey {
-  static var liveValue: DatabaseQueue {
-    let path = URL.documentsDirectory.appending(component: "db.sqlite").path()
-    print("open", path)
-    var configuration = Configuration()
-    configuration.prepareDatabase { db in
-      db.trace { event in
-        print(event)
-      }
-    }
-    let databaseQueue = try! DatabaseQueue(path: path, configuration: configuration)
-    try! databaseQueue.migrate()
-    return databaseQueue
-  }
-
-  static var previewValue: DatabaseQueue {
-    testValue
-  }
-
-  static var testValue: DatabaseQueue {
-    let databaseQueue = try! DatabaseQueue()
-    try! databaseQueue.migrate()
-    return databaseQueue
   }
 }
